+++ src/main/java/com/soklet/Soklet.java
@@ -2634,12 +2634,6 @@ public final class Soklet implements AutoCloseable {
 			return getRealImplementation().getSessionStore();
 		}
 
-			@NonNull
-			@Override
-			public IdGenerator<String> getSessionIdGenerator() {
-				return getRealImplementation().getSessionIdGenerator();
-			}
-
 		@NonNull
 		protected McpServer getRealImplementation() {
 			return this.realImplementation;